reconfigured project to not need home.js and made ui changes
